requirements.txt
asgiref==3.7.2
attrs==23.1.0
bottle==0.12.25
cffi==1.15.1
clr-loader==0.2.6
Django==4.2.5
pendulum==2.1.2
proxy-tools==0.1.0
pycparser==2.21
python-dateutil==2.8.2
pythonnet==3.0.2
pytzdata==2020.1
pywebview==4.3.2
six==1.16.0
sqlparse==0.4.4
supermemo2==2.0.0
typing_extensions==4.7.1
tzdata==2023.3
